refactor: split orchestration persistence bridge
