+++ crates/tandem-server/src/app/state/app_state_impl_parts/part01.rs
@@ -491,22 +491,7 @@ impl AppState {
     }
 
     pub async fn mark_ready(&self, runtime: RuntimeState) -> anyhow::Result<()> {
-        {
-            let mut guard = self
-                .stateful_engine_lock
-                .lock()
-                .map_err(|_| anyhow::anyhow!("stateful engine lock guard was poisoned"))?;
-            if guard.is_none() {
-                let paths =
-                    crate::stateful_runtime::OrchestrationStorePaths::from_automation_runs_path(
-                        &self.automation_v2_runs_path,
-                    );
-                let engine_lock =
-                    crate::stateful_runtime::StatefulEngineLock::acquire(&paths.engine_lock_path)?;
-                let _store = crate::stateful_runtime::OrchestrationStateStore::open(paths)?;
-                *guard = Some(engine_lock);
-            }
-        }
+        self.acquire_stateful_engine_lock()?;
         runtime
             .engine_loop
             .set_tool_dispatch_ledger(Arc::new(AppStateToolDispatchLedger {
@@ -1686,15 +1671,7 @@ impl AppState {
 
     pub async fn load_automation_v2_runs(&self) -> anyhow::Result<()> {
         let mut merged = std::collections::HashMap::<String, AutomationV2RunRecord>::new();
-        let automation_runs_path = self.automation_v2_runs_path.clone();
-        let database_runs = tokio::task::spawn_blocking(move || {
-            crate::stateful_runtime::OrchestrationStateStore::from_automation_runs_path(
-                &automation_runs_path,
-            )?
-            .load_automation_runs()
-        })
-        .await
-        .map_err(|error| anyhow::anyhow!("automation run database load task failed: {error}"))??;
+        let database_runs = self.load_automation_v2_runs_from_stateful_store().await?;
         let mut loaded_from_alternate = false;
         let mut canonical_loaded = false;
         let mut runs_store_upgraded = false;
@@ -1783,20 +1760,8 @@ impl AppState {
             dropped_nonterminal_recovered_context_runs,
             "loaded automation v2 runs"
         );
-        let database_snapshot = merged.values().cloned().collect::<Vec<_>>();
-        let automation_runs_path = self.automation_v2_runs_path.clone();
-        let imported_at_ms = now_ms();
-        tokio::task::spawn_blocking(move || {
-            let store =
-                crate::stateful_runtime::OrchestrationStateStore::from_automation_runs_path(
-                    &automation_runs_path,
-                )?;
-            store.import_legacy_runs(&automation_runs_path, &database_snapshot, imported_at_ms)
-        })
-        .await
-        .map_err(|error| {
-            anyhow::anyhow!("automation run database import task failed: {error}")
-        })??;
+        self.import_automation_v2_runs_to_stateful_store(&merged)
+            .await?;
         *self.automation_v2_runs.write().await = merged;
         let recovered = self
             .recover_automation_definitions_from_run_snapshots()
@@ -1824,43 +1789,6 @@ impl AppState {
             let _ =
                 cleanup_stale_legacy_automation_v2_runs_file(&self.automation_v2_runs_path).await;
         }
-        Ok(())
-    }
-
-    pub async fn persist_automation_v2_runs(&self) -> anyhow::Result<()> {
-        let (runs_snapshot, automations_snapshot) = {
-            let runs = self.automation_v2_runs.read().await;
-            let automations = self.automations_v2.read().await;
-            (runs.clone(), automations.clone())
-        };
-        let database_snapshot = runs_snapshot.values().cloned().collect::<Vec<_>>();
-        let automation_runs_path = self.automation_v2_runs_path.clone();
-        tokio::task::spawn_blocking(move || {
-            crate::stateful_runtime::OrchestrationStateStore::from_automation_runs_path(
-                &automation_runs_path,
-            )?
-            .upsert_automation_runs(database_snapshot.iter())
-        })
-        .await
-        .map_err(|error| {
-            anyhow::anyhow!("automation run database persist task failed: {error}")
-        })??;
-        for run in runs_snapshot.values() {
-            write_automation_v2_run_history_shard(&self.automation_v2_runs_path, run).await?;
-        }
-        let mut compacted = runs_snapshot;
-        compacted.retain(|_, run| !automation_v2_run_is_nonterminal_recovered_context_run(run));
-        compact_automation_v2_runs_for_hot_storage(
-            &mut compacted,
-            &automations_snapshot,
-            automation_v2_hot_cutoff_ms(),
-        );
-        let payload = serialize_automation_v2_runs_file(compacted)?;
-        if let Some(parent) = self.automation_v2_runs_path.parent() {
-            fs::create_dir_all(parent).await?;
-        }
-        fs::write(&self.automation_v2_runs_path, &payload).await?;
-        let _ = cleanup_stale_legacy_automation_v2_runs_file(&self.automation_v2_runs_path).await;
         Ok(())
     }
